Refactor field nulling.
